Tweak NumericFormat.digitsOf()
Make NumericFormat.digitsOf() return List[Int] instead of
List[Long].

+++ src/main/scala/io/github/nthportal/euler/h0/t4/Problem40.scala
@@ -8,7 +8,7 @@ object Problem40 extends ProjectEulerProblem {
     droppingStream(Stream.from(1).flatMap(NumericFormat.digitsOf), (0 to 6).map(10 ** _).map(_.toInt).toStream).product
   }
 
-  private def droppingStream(stream: Stream[Long], targets: Stream[Int], alreadyDropped: Int = 0): Stream[Long] = {
+  private def droppingStream(stream: Stream[Int], targets: Stream[Int], alreadyDropped: Int = 0): Stream[Long] = {
     if (targets.isEmpty) {
       Stream.empty
     } else {
